// src/app/model/Employee.ts
export class Employee{
   empId: number;
   name: string;
   city: string;
   state: string;
   email: string;
   contact: string;
   adresse: string;
   pinCode: string;
   constructor(){
    this.name='';
    this.city='';
    this.state= '';
    this.email='';
    this.contact='';
    this.adresse='';
    this.empId=1;
    this.pinCode= '';
   }
}